reorganize comparer singleton
