+++ src/NMoneys/NMoneys.csproj
@@ -109,7 +109,6 @@
     <Compile Include="Money.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Support\Extensions.cs" />
-    <Compile Include="Support\FastEnumComparer.cs" />
     <Compile Include="Support\Guard.cs" />
     <Compile Include="Support\Range.cs" />
     <Compile Include="Support\TokenizedValue.cs" />
